+++ db/sql/see-user-settings.sql
@@ -1,0 +1,1 @@
+SELECT * FROM wp_testing_4_1.wp_usermeta;